Refactor jar resource loading out to BetterUtils::getResourceFromJar
This fixes an issue where the built jar would crash upon loading, because this::class.java.classLoader would be null because it was loaded with the BootstrappedClassLoader. By having a convenient new method for this, we can hope that I, or anyone else, doesn't fall prey to the same mistake again, and another five hours of debugging to end up having to change `this` to `Minecraft` and be done.